+++ src/components/common/member-button.tsx
@@ -14,6 +14,8 @@ export const MemberButton = ({
     <Link
       href={url}
       className="flex h-14 items-center justify-start gap-4 rounded-lg bg-gradient-to-b from-neutral-700 to-neutral-800 px-4"
+      target="_blank"
+      rel="noopener noreferrer"
     >
       <Image
         src={imgLink}
