+++ src/server/api/routers/about.ts
@@ -1,19 +1,16 @@
-import { type GithubMemberProps } from "@/components/about/member-section";
+import { type GithubMemberProps } from "@/types/about";
 import { createTRPCRouter, publicProcedure } from "@/server/api/trpc";
 // import { TRPCError } from "@trpc/server";
 
-export const aboutMemberRouter = createTRPCRouter({
+export const aboutRouter = createTRPCRouter({
   get: publicProcedure.query(async () => {
     const response = await fetch(
       `https://api.github.com/orgs/esc-chula/public_members?per_page=200`,
     );
 
-    // if (!response.ok) {
-    //   throw new TRPCError({
-    //     code: 'INTERNAL_SERVER_ERROR',
-    //     message: 'Failed to fetch GitHub members',
-    //   });
-    // }
+    if (!response.ok) {
+      throw new Error("Can't get ESC members");
+    }
 
     const data: GithubMemberProps[] =
       (await response.json()) as GithubMemberProps[];
